// src/app/crmadmin/(services)/visastatus/types/IVisaStatus.tsx
export interface VisaStatusResponse {
    id: string
    name: string
    visaStatusType: number

}

export interface AddVisaStatusResponse {
    id: string
    name: string
    visaStatusType: number
}


export interface UpdateVisaStatusResponse {
    id: string
    name: string
    visaStatusType: number
}

export interface UpdateVisaStatusPayload {
    name: string
    visaStatusType: number

}

export interface AddVisaStatusPayload {
    name: string
    visaStatusType: number

}